Added jacket artwork searching
If Shazam can pull back a match, the album artwork for that song will be parsed in. You can click the little album button that appears next to the play button and you can compare the artwork and choose to update or not. The new artwork will be resized to be the same as the old artwork.