Bug fixes and rainbows
Didn't realize there were so many bugs I added (squash one and 10 pop up).
Added rainbows because it needs rainbows

There was a big issue with sorting -- if you sorted by a column and edited a value in that column it would immediately sort to the row it belonged it.  Updated to turn off sorting right after a sort happens to prevent this.